Rename Schedule container to Schedule
