+++ src/App.js
@@ -7,7 +7,7 @@ import {
   faArrowCircleLeft,
 } from '@fortawesome/free-solid-svg-icons'
 import Home from './Home'
-import ScheduleContainer from './ScheduleContainer'
+import Schedule from './Schedule'
 import {HashRouter, Switch, Route} from 'react-router-dom'
 
 library.add(faBus, faHome, faArrowCircleLeft)
@@ -19,7 +19,7 @@ class App extends Component {
         <div className="flex flex-column h-100 pa3 bg-near-black items-center">
           <Switch>
             <Route exact path="/" component={Home} />
-            <Route path="/schedule/:line/:stop" component={ScheduleContainer} />
+            <Route path="/schedule/:line/:stop" component={Schedule} />
           </Switch>
         </div>
       </HashRouter>
